Refactor filmstrip resize calculations
Simplify and remove unnecessary calls to recompute.
